feat: 🎸 Add popup
