+++ src/app/src/pages/Popup/popup.module.styl
@@ -1,0 +1,89 @@
+.warp {
+  display: flex
+  justify-content: center
+  align-items: center
+  min-width: 180px
+  min-height: 130px
+  font-size: 14px
+}
+
+.name {
+  display: flex
+  align-items: center
+  padding: 0 8px + 4px
+  height: 37px
+
+  h1 {
+    font-size: 14px
+
+    a {
+      color: #333
+    }
+  }
+}
+
+.small {
+  color: #ccc
+  font-size: 12px
+
+  a {
+    color: #ccc
+
+    &:hover {
+      text-decoration: underline
+    }
+  }
+}
+
+.nav {
+  margin: 0
+  padding: 0
+  list-style: none
+
+  .item {
+    display: block
+    box-sizing: border-box
+    margin: 2px 0
+  }
+
+  .item > a > .glyphicon {
+    margin-right: 8px
+  }
+
+  .item > a {
+    display: block
+    margin: 0 5px
+    padding: 8px 25px 8px 8px
+    border-radius: 4px
+    color: #337ab7
+    text-decoration: none
+    white-space: nowrap
+    font-size: 1em
+    line-height: 1.5em
+    cursor: pointer
+  }
+
+  .item > a:hover {
+    background-color: #eee
+    color: #23527c
+    text-decoration: none
+  }
+}
+
+.hr {
+  // overflow: hidden
+  // height: 1px
+  // background-color: #E5E5E5
+  border: none
+  border-bottom: 0.5px solid #dcdcdc
+}
+
+.footer {
+  display: flex
+  justify-content: flex-end
+  align-items: center
+  padding: 0 8px
+  height: 37px
+  color: #999
+  font-size: 12px
+}
