+++ src/background/context_menus.ts
@@ -1,56 +1,4 @@
-import { genIssueUrl, isChinese } from '@@/utils'
-import { openPage } from './utils'
-// const repositoryUrl: string = repository.url
-
-type Contexts = chrome.contextMenus.ContextType[]
-interface IInitContextMenu {
-  title: string
-  contexts: Contexts
-  onclick: (info: chrome.contextMenus.OnClickData, tab: chrome.tabs.Tab | undefined) => void
-}
-
-const contextMenus: Record<string, IInitContextMenu> = {
-  // version: {
-  //   title: `🧃 Version: ${version}`,
-  //   contexts: ['action'],
-  //   onclick: () => {
-  //     openPage(`${repositoryUrl}/releases/tag/${version}`)
-  //   }
-  // },
-  openChat: {
-    title: '💬 New Bing',
-    contexts: ['action'],
-    onclick: (_info) => {
-      openPage('https://www.bing.com/search?q=Bing+AI&showconv=1')
-    }
-  },
-
-  openImageCreate: {
-    title: '🖼️ New Bing Image Creator',
-    contexts: ['action'],
-    onclick: (_info) => {
-      openPage('https://www.bing.com/create')
-    }
-  },
-
-  likeIt: {
-    title: '❤️ Like it',
-    contexts: ['action'],
-    onclick: () => {
-      openPage('https://chrome.google.com/webstore/detail/new-bing-anywhere-bing-ch/hceobhjokpdbogjkplmfjeomkeckkngi/reviews')
-    }
-  },
-
-  reportIssues: {
-    title: isChinese ? '🐛 反馈建议' : '🐛 Report issues',
-    contexts: ['action'],
-    onclick: async (_info) => {
-      const url = await genIssueUrl()
-
-      openPage(url)
-    }
-  }
-}
+import { contextMenus } from '@@/utils'
 
 export default () => {
   chrome.contextMenus.removeAll(() => {
