+++ src/background/utils.ts
@@ -43,49 +43,6 @@ export const dumpTabs = async ({ windowId }): Promise<void> => {
   })
 }
 
-export const getURL = (url: string = '', base?: string): URL => {
-  try {
-    return new URL(url, base)
-  } catch (e) {
-    // console.error(e)
-    return {
-      searchParams: {
-        get: () => null
-      }
-    } as any
-  }
-}
-
-export const getURLSearchParams = (url: string): URLSearchParams => {
-  try {
-    return new URLSearchParams(url)
-  } catch {
-    return {
-      get: () => null
-    } as any
-  }
-}
-
-export const openPage = async (url: string): Promise<chrome.tabs.Tab> => {
-  const tabs = await chrome.tabs.query({ currentWindow: true })
-
-  const urlObj = getURL(url)
-  let tab = tabs.find((tab) => tab.url?.startsWith(urlObj.origin))
-
-  if (tab == null) {
-    tab = await chrome.tabs.create({ url })
-  } else {
-    await Promise.all(
-      [
-        chrome.tabs.move(tab.id!, { index: tabs.length - 1 }),
-        tab.url !== url && chrome.tabs.update(tab.id!, { url }),
-        chrome.tabs.update(tab.id!, { active: true, url: tab.url !== url ? url : undefined })
-      ].filter(Boolean)
-    )
-  }
-  return tab
-}
-
 export const setCookie = async (options: chrome.cookies.SetDetails, cookie: chrome.cookies.Cookie = {} as any) => {
   return await chrome.cookies.set({
     domain: cookie.domain,
